fixed glitchy player
dont use irational numbers for rotations on the platforms use numbers like 0.13
please no greater than 3 SF

+++ Gradded-Unit/src/com/FuturePixels/Drawables/Levels/PlatForm.java
@@ -24,7 +24,7 @@ public class PlatForm extends IDrawable {
         super();
     }
 
-    public PlatForm(Vector position, float RadianRotation) {
+    public PlatForm(Vector position, double RadianRotation) {
         super();
         GetSprite("/images/Platform.png");
         setPosition(position);
